Controle de saisie et photo de profile

+++ Front/src/app/views/profil/profil.component.ts
@@ -1,7 +1,9 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, NgForm } from '@angular/forms';
 import { Router } from '@angular/router';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { FormControl } from '@angular/forms';
 import {
   RowComponent, ColComponent, CardComponent, CardHeaderComponent,
   CardBodyComponent, FormControlDirective, FormDirective,
@@ -27,6 +29,7 @@ import { GererUtilisateurService } from '../utilisateurs/utilisateurs-basic-exam
   styleUrls: ['./profil.component.scss']
 })
 export class profil {
+  userForm: FormGroup;
   user: any;
 
   // Options pour les selects
@@ -35,7 +38,24 @@ export class profil {
   formationOptions = ['INGENIERIE', 'MASTERE'];
   sexeoptions = ['HOMME', 'FEMME'];
 
-  constructor(private GererUtilisateurService: GererUtilisateurService, private router: Router) { }
+  constructor(private fb: FormBuilder, private GererUtilisateurService: GererUtilisateurService, private router: Router) {
+    this.userForm = this.fb.group({
+      nom: ['', [Validators.required, Validators.minLength(2)]],
+      prenom: ['', [Validators.required, Validators.minLength(2)]],
+      cin: ['', [Validators.pattern('[0-9]{8}'), Validators.required]],
+      sexe: ['', Validators.required],
+      dateDeNaissance: ['', Validators.required],
+      lieuNaissance: [''],
+      email: ['', [Validators.required, Validators.email]],
+      tel: ['', [Validators.pattern('[0-9]{8,15}'), Validators.required]],
+      adresse: [''],
+      fax: ['', Validators.pattern('[0-9]+')],
+      filiere: ['', Validators.required],
+      niveau: ['', Validators.required],
+      formation: ['', Validators.required],
+      option: ['']
+    });
+  }
   ngOnInit(): void {
     this.loadCurrentUser();
   }
@@ -64,38 +84,43 @@ export class profil {
     }
   }
 
-  onSubmit(): void {
+  onSubmit(form: NgForm): void {
+    if (form.valid) {
 
 
-    Swal.fire({
-      title: 'Êtes-vous sûr ?',
-      text: 'Voulez vous modifier ces informations',
-      icon: 'warning',
-      showCancelButton: true,
-      confirmButtonText: 'Oui, Modifier!',
-      cancelButtonText: 'Annuler',
-      reverseButtons: true
-    }).then((result) => {
-      if (result.isConfirmed) {
-        // L'événement est confirmé
-        this.GererUtilisateurService.updateProfile(this.user).subscribe(
-          (response) => {
-            console.log('Donnees modifiée avec succès', response);
-            localStorage.setItem('user', JSON.stringify(this.user));
-            window.location.reload();
-          },
-          (error) => {
-            console.error('Erreur lors de la modification', error);
-          }
-        );
-        // Ajouter la logique de confirmation ici
-      } else if (result.isDismissed) {
-        // L'événement est annulé
-        Swal.fire('Événement annulé', '', 'info');
-      }
-    });
-    // Ici vous ajouteriez l'appel à votre service
-    // this.userService.updateUser(this.user).subscribe(...)
+      Swal.fire({
+        title: 'Êtes-vous sûr ?',
+        text: 'Voulez vous modifier ces informations',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Oui, Modifier!',
+        cancelButtonText: 'Annuler',
+        reverseButtons: true
+      }).then((result) => {
+        if (result.isConfirmed) {
+          // L'événement est confirmé
+          console.log(this.user)
+          this.GererUtilisateurService.updateProfile(this.user).subscribe(
+            (response) => {
+              console.log('Donnees modifiée avec succès', response);
+              localStorage.setItem('user', JSON.stringify(this.user));
+              window.location.reload();
+            },
+            (error) => {
+              console.error('Erreur lors de la modification', error);
+            }
+          );
+          // Ajouter la logique de confirmation ici
+        } else if (result.isDismissed) {
+          // L'événement est annulé
+          Swal.fire('Événement annulé', '', 'info');
+        }
+      });
+      // Ici vous ajouteriez l'appel à votre service
+      // this.userService.updateUser(this.user).subscribe(...)
+    } else {
+      console.log("Le formulaire est invalide");
+    }
   }
 
   onCancel(): void {
@@ -120,4 +145,6 @@ export class profil {
       }
     });
   }
+
 }
+
